moved properties that were in Player.cs to Projectile.cs and fixed the collision in collectionController.cs appropriately also Dagger.cs is now Projectile.cs I tested things out and fixed a few things. Eveything seems to be working but there could be a few bugs still.
