+++ Assets/Resources/Prefabs/ItemPrefabs/ThrowingDagger_Item.prefab
@@ -124,20 +124,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5cf5a9889d81083489640954855f5316, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  filter:
-    useTriggers: 0
-    useLayerMask: 0
-    useDepth: 0
-    useOutsideDepth: 0
-    useNormalAngle: 0
-    useOutsideNormalAngle: 0
-    layerMask:
-      serializedVersion: 2
-      m_Bits: 0
-    minDepth: 0
-    maxDepth: 0
-    minNormalAngle: 0
-    maxNormalAngle: 0
   name: Boot
   description: Give the player +1 movespeed
   itemSprite: {fileID: 6440375, guid: 88b39d6be2296fb42b2e24e967dd0b77, type: 3}
@@ -146,7 +132,12 @@ MonoBehaviour:
   moveSpeed: 1
   attackSpeed: 0
   pushRecovery: 0
-  weapon: {fileID: 0}
-  projectile: {fileID: 0}
+  weapon: {fileID: 3334614298463244672, guid: a1a252e1299c3fd4e86d3e5ce0739511, type: 3}
+  projectile: {fileID: 2120853311608965495, guid: 2502f902d8f0b8f45bfdcaf484dfc2f2, type: 3}
   trans: {fileID: 0}
   lifespan: 0
+  critChance: 0
+  critMultiplier: 0
+  playerDamage: 0
+  lifesteal: 0
+  coll: {fileID: 0}
