--- Assets/Scripts/Dagger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20daf12853904ab489d0227c0280325e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
